Move the Sampler interface to the sdk
