+++ src/api/trace/tracer.rs
@@ -150,6 +150,7 @@ use crate::api::{
     context::{Context, ContextGuard},
     TraceContextExt,
 };
+use crate::sdk;
 use std::fmt;
 use std::time::SystemTime;
 
@@ -422,7 +423,7 @@ pub struct SpanBuilder {
     /// Span status message
     pub status_message: Option<String>,
     /// Sampling result
-    pub sampling_result: Option<api::SamplingResult>,
+    pub sampling_result: Option<sdk::SamplingResult>,
 }
 
 /// SpanBuilder methods
@@ -535,7 +536,7 @@ impl SpanBuilder {
     }
 
     /// Assign sampling result
-    pub fn with_sampling_result(self, sampling_result: api::SamplingResult) -> Self {
+    pub fn with_sampling_result(self, sampling_result: sdk::SamplingResult) -> Self {
         SpanBuilder {
             sampling_result: Some(sampling_result),
             ..self
